findinlisppath.lsp: simple routine to find a file in *lisp-path*
My Lisp/Scheme is a little rusty (it's been 20 years, and I was never
goood at it), but I thought there should be a mechanism to pick up
per-user config (if it exists), and falling back on system-wide files
otherwise.

To this end, I have a global variable, *lisp-path*, which defaults to

    ~/.config/Femto-Emacs
    ~
    *install-dir*/../share/Femto-Emacs
    *install-dir*

where *install-dir* is the directory the binary is in.

Rationale: ~/.config/Femto-Emacs is the "tidy" option for user
custimization, since it keeps your home directory relatively clean. ~
(the user's home directory) is second, to keeps things working for
existing users who have the customization in ~. If your femto is
installed in /usr/local/bin (or /opt/femto/bin), the third entry will
resolve to /usr/local/share/Femto-Emacs (or
/opt/femto/share/Femto-Emacs), so will most likely do the right thing
for people who install Femto-Emacs system-wide, but don't want copy the
initialisation files for every user. Lastly, *install-dir* is meant to
help pick up the files when you've just built Femto-Emacs, and just want
to take it for a quick spin (admittedly, this last element has the
weakest justification).

There is also a find-in-lisp-path routine which finds its argument in
the *lisp-path* (first match wins), or returns #f if unsuccessful.